% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_calculations.R
\name{calculate_annual_costs}
\alias{calculate_annual_costs}
\title{Calculate Annual Costs by Person}
\usage{
calculate_annual_costs(cost_data)
}
\arguments{
\item{cost_data}{Cost data from get_*_costs functions}
}
\value{
Tibble with annual costs per person
}
\description{
Calculate Annual Costs by Person
}
