% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_calculations.R
\name{calculate_cost_summary}
\alias{calculate_cost_summary}
\title{Calculate Summary Statistics}
\usage{
calculate_cost_summary(cost_data, group_vars = "cohort")
}
\arguments{
\item{cost_data}{Cost data tibble from calculate_annual_costs function}

\item{group_vars}{Character vector of variables to group by}
}
\value{
Tibble with summary statistics
}
\description{
Calculate Summary Statistics
}
