% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_functions.R
\name{plot_cost_trends}
\alias{plot_cost_trends}
\title{Plot Cost Trends by Type or Patient Number}
\usage{
plot_cost_trends(types_data, cost_var = "total_cost")
}
\arguments{
\item{types_data}{Data from prepare_types_data()}

\item{cost_var}{Variable to plot: total_cost/n_patients (default: "total_cost")}
}
\value{
ggplot object
}
\description{
Line chart showing cost trends over time by type or number of patients over time
}
