% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_functions.R
\name{plot_costs_by_type}
\alias{plot_costs_by_type}
\title{Plot Costs by Type}
\usage{
plot_costs_by_type(type_graph_data)
}
\arguments{
\item{type_graph_data}{Data from prepare_type_graph_data()}
}
\value{
ggplot object
}
\description{
Plot Costs by Type
}
