% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_functions.R
\name{plot_costs_by_type_year}
\alias{plot_costs_by_type_year}
\title{Plot Costs by Type and Year}
\usage{
plot_costs_by_type_year(types_data)
}
\arguments{
\item{types_data}{Data from prepare_types_data()}
}
\value{
ggplot object
}
\description{
Plot Costs by Type and Year
}
