% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_calculations.R
\name{prepare_graph_data}
\alias{prepare_graph_data}
\title{Prepare Data for Analysis by Year}
\usage{
prepare_graph_data(costs)
}
\arguments{
\item{costs}{Cost data from get_*_costs()}
}
\value{
Tibble ready for visualization
}
\description{
Creates aggregated summary by year, and cohort
}
