% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_calculations.R
\name{prepare_types_data}
\alias{prepare_types_data}
\title{Prepare Data for Shiny Application}
\usage{
prepare_types_data(all_costs)
}
\arguments{
\item{all_costs}{Combined cost data from get_all_costs()}
}
\value{
List
}
\description{
Creates 3 tables: aggregated summary by cost type, year, and cohort; person-level costs with and without type column
}
