% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_functions.R
\name{plot_annual_costs}
\alias{plot_annual_costs}
\title{Plot Annual Costs Over Time}
\usage{
plot_annual_costs(types_data)
}
\arguments{
\item{types_data}{Data from prepare_types_data()}
}
\value{
ggplot object
}
\description{
Plot Annual Costs Over Time
}
